requirements.txt
# Core packages
yfinance<=0.2.55
pandas>=1.3.0
autogen>=0.8.5
numpy>=1.20.0
matplotlib>=3.5.0
yfinance>=0.2.55
pyautogen>=0.8.5

# LLM providers
openai>=1.64.0
google-generativeai>=0.3.0
groq>=0.4.0

# Data visualization
streamlit>=1.44.1
plotly>=5.13.0
seaborn>=0.12.0

# Data analysis
pandas-datareader>=0.10.0
statsmodels>=0.13.0

# Development tools
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.0

# Network and storage
requests>=2.28.0
diskcache>=5.6.0

# AI/ML libraries
scikit-learn>=1.0.0
prophet>=1.0.0

# SSL handling
certifi>=2025.1.31

# Add python-dotenv to existing requirements
python-dotenv>=1.0.0

tensorflow
vertexai
ag2
jsonschema
ag2[gemini]


